requirements-web.txt
pygbag>=0.9.2
pygame-ce>=2.5
